unica: rro: don't add overlay if target=source

+++ unica/patches/rro/customize.sh
@@ -1,9 +1,11 @@
-if [[ -d "$SRC_DIR/target/$TARGET_CODENAME/overlay" ]]; then
-    DECODE_APK "product" "overlay/framework-res__a73xqxx__auto_generated_rro_product.apk"
+if [[ $TARGET_FIRMWARE != $SOURCE_FIRMWARE ]]; then
+    if [[ -d "$SRC_DIR/target/$TARGET_CODENAME/overlay" ]]; then
+        DECODE_APK "product" "overlay/framework-res__a73xqxx__auto_generated_rro_product.apk"
 
-    echo "Applying stock overlay configs"
-    rm -rf "$APKTOOL_DIR/product/overlay/framework-res__a73xqxx__auto_generated_rro_product.apk/res"
-    cp -a --preserve=all \
-        "$SRC_DIR/target/$TARGET_CODENAME/overlay" \
-        "$APKTOOL_DIR/product/overlay/framework-res__a73xqxx__auto_generated_rro_product.apk/res"
+        echo "Applying stock overlay configs"
+        rm -rf "$APKTOOL_DIR/product/overlay/framework-res__a73xqxx__auto_generated_rro_product.apk/res"
+        cp -a --preserve=all \
+            "$SRC_DIR/target/$TARGET_CODENAME/overlay" \
+            "$APKTOOL_DIR/product/overlay/framework-res__a73xqxx__auto_generated_rro_product.apk/res"
+    fi
 fi
